refactor(formats): convert gene_model into package
